backend/requirements.txt
fastapi==0.115.11
loguru==0.7.3
pydantic-settings==2.8.1
python-dotenv==1.0.1
python-multipart==0.0.20
sqlmodel==0.0.24
typer==0.15.2
uvicorn==0.34.0
PyJWT==2.10.1
user-agents==2.2.0
passlib==1.7.4
bcrypt==4.0.1 